chore_: move and rename server files
