+++ cmd/status-backend/main.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/ethereum/go-ethereum/log"
 
-	"github.com/status-im/status-go/cmd/statusd/server"
 	"github.com/status-im/status-go/logutils"
+	"github.com/status-im/status-go/cmd/status-backend/server"
 )
 
 var (
