+++ cmd/status-backend/server/parse-api/endpoints_template.txt
